Move variations to block js
