--- src/App/Blocks/src/variations/index.js
@@ -1,1 +1,0 @@
-import './eventsByTag';
